Dev2

+++ src/main/kotlin/ar/edu/unsam/phm/uberto/dto/TripScoreDTO.kt
@@ -1,7 +1,7 @@
 package ar.edu.unsam.phm.uberto.dto
 
 data class TripScoreDTO(
-    val tripId: Int,
+    val tripId: Long,
     val message: String,
     val scorePoints: Int,
     val date: String,
@@ -11,13 +11,3 @@ data class TripScoreDTO(
     val avatarUrlDriver: String,
     val delete: Boolean
 )
-
-
-//fun TripScore.toDTO() = TripScoreDTO(
-//    message = message,
-//    scorePoints = scorePoints,
-//    date = date.toString(),
-//    passengerId= passenger.userId,
-//    avatarUrl = passenger.img
-//)
-
